Renamed projects Downgrade the .net version to 4.0 for the core library
