--- src/EventStore.Tools.Infrastructure.Repository/EventStore.Tools.Infrastructure.Repository.nuspec
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>EventStore.Tools.Infrastructure.Repository</id>
-    <version>1.1.0.0</version>
-    <title>EventStore.Tools.Infrastructure.Repository</title>
-    <authors>Riccardo Di Nuzzo</authors>
-    <owners>Riccardo Di Nuzzo</owners>
-    <licenseUrl>https://github.com/riccardone/EventStore.Tools.Infrastructure/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/riccardone/EventStore.Tools.Infrastructure.Repository</projectUrl>    
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>This library contains an EventStoreDomainRepository based on a DomainRepository base class contained in EventStore.Tools.Infrastructure.Repository library. The two libraries combined are used to build Event Sourced components interacting with EventStore (GetEventStore)</description>
-    <releaseNotes>Removed the correlationId from the Save IDomainRepository method and use AggregateId instead</releaseNotes>
-    <copyright>Copyright 2017</copyright>
-    <tags>GetEventStore, EventStore, DomainRepository, Repository, DDD</tags>
-  </metadata>
-</package>
