+++ src/Evento.Repository/EventStoreDomainRepository.cs
@@ -5,7 +5,7 @@ using System.Text;
 using EventStore.ClientAPI;
 using Newtonsoft.Json;
 
-namespace EventStore.Tools.Infrastructure.Repository
+namespace Evento.Repository
 {
     public class EventStoreDomainRepository : DomainRepositoryBase
     {
