+++ src/Evento.Repository/packages.config
