+++ src/Evento/DomainRepositoryBase.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace EventStore.Tools.Infrastructure
+namespace Evento
 {
     public abstract class DomainRepositoryBase : IDomainRepository
     {
