+++ src/Infrastructure.Tests/Fakes/FakeAggregate.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using EventStore.Tools.Infrastructure;
+﻿using Evento;
 
 namespace Infrastructure.Tests.Fakes
 {
